proper handling of composite indexes in the automatic update handler
for composite indexes we make sure to first check if the resource can be linked
to any of the related types and only if it relates to none do we proceed in
deleting the document.

if it matches a type, we figure out which subindexes are matches and build
documents for each of them. In case of multiple matches we merge those documents
into one document. This strategy seems to be the most sane since it is
impossible for a user of the service to determine which subindex would "win" if
we would only retain one.