fix delta handling wrt to deletes
the goal of the previous logic was to optimize handling of delta's, but it made
some incorrect assumptions.

1. it's not because we receive a delete and insert in one changeset that this
should be considered as an update. it could very well be that data was deletes
as the result of this changeset
2. for finding affected documents of a delta we do need to know whether a delta
is a delete or an insert: for deletes we can only construct the path from the
root of the document to the subject of a the deleted triple, but obviously not
to the object of the triple (since this would no longer be there and not be
found)

this commit fixes both assumptions, by re-enabling the different handling for
inserts and deletes