support large yaml files
fixes #92. This bug was introduced in beta.5 as we snakeyaml was upgraded in
jruby to [1.32](https://bitbucket.org/snakeyaml/snakeyaml/wiki/Changes).
This release introduced a safe file size limit of 3MB. We currently overwrite this default and set up a default of 20MB, which we may want to make configurable later on.